[update] cierre implementación layer
