+++ lambda-layers/common/python/lambda_utils/__init__.py
@@ -2,4 +2,4 @@ from .decorators import lambda_handler_decorator
 
 __all__ = [
     "lambda_handler_decorator"
-    ]
+]
