--- src/delete/requirements.txt
@@ -1,1 +1,0 @@
-aws-lambda-powertools
