--- src/get/get_by_pk/requirements.txt
@@ -1,1 +1,0 @@
-aws-lambda-powertools
